\name{assess.doneness}
\alias{assess.doneness}
\title{assess doneness}
\description{Do we have a big enough call-vs-noise sample size to be able to make a
model which will save us labor overall?}
\usage{assess.doneness(bootstrap)}
\arguments{
  \item{bootstrap}{The bootstrap call-vs-noise data set. Features from generate.seewave.measures() must be present, as well as a Call_vs_Noise column
filled in with "Call" and "Noise".}
}

\value{Currently no return value; the doneness must be assessed by visual 
evaluation of the plots generated by this function.}

\author{Jesse C. Ross}



\seealso{\code{\link{nfold.xval}}}

