\name{flightcallr-package}
\alias{flightcallr-package}
\alias{flightcallr}
\docType{package}
\title{Classify Night Flight Calls Based on Acoustic Measurements}
\description{flightcallr uses the seewave package to generate acoustic
measurements ("features") of sounds, and then trains and uses random
forest models to make predictions based on these features. Accuracy
can be constrained via cross-validation. Convenience functions for
reading and writing Raven selection tables are included.}
\details{
\tabular{ll}{Package: \tab flightcallr\cr
Type: \tab Package\cr
Title: \tab Classify Night Flight Calls Based on Acoustic Measurements\cr
Version: \tab 0.04\cr
Date: \tab 2013-07-30\cr
Author: \tab Jesse C. Ross\cr
Depends: \tab seewave, tuneR, randomForest, ROCR, R (>= 2.10)\cr
Maintainer: \tab Jesse C. Ross <jross@cornell.edu>\cr
License: \tab GPL-3\cr}
}
\author{Jesse C. Ross}

\keyword{ package }


