% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nfold.xval}
\alias{nfold.xval}
\title{Perform n-fold cross-validation to generate a perf object.}
\usage{
nfold.xval(measurements, annotation = "", fold = 10, plot.ROC = T,
  color.plot = T)
}
\arguments{
\item{measurements}{A set of selection table rows with Call_vs_Noise column
filled in, and with features as generated by generate.seewave.measures()}

\item{annotation}{An optional annotation for the plots.}

\item{fold}{n, i.e. how many rounds of cross-validation to perform.}

\item{plot.ROC}{Whether or not to plot the ROC curve pertaining to the
cross-validation run.}

\item{color.plot}{Whether or not the plot should be in color.}
}
\value{
\item{name perf}{description a perf object}
        \item{name difficulty}{description a vector signifying how
          frequently the row in question is incorrectly identified in
          cross-validation}
}
\description{
Perform n-fold cross-validation to generate a perf object.
}

