% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rejigger.perf}
\alias{rejigger.perf}
\title{Rejigger a perf object from ROCR so that it is no longer sparse.}
\usage{
rejigger.perf(perf)
}
\arguments{
\item{perf}{The perf object to be rejiggered.}
}
\value{
The non-sparse representation of that perf object.
}
\description{
Rejigger a perf object from ROCR so that it is no longer sparse.
}

