\name{seewave.measures}
\alias{seewave.measures}
\docType{data}
\title{List of seewave measurements returned by generate.seewave.measurements()}
\description{
This is just a list of the names of the seewave measurements which are added to a selection table by \link{generate.seewave.measures}.
}
\usage{data(seewave.measures)}
\format{
  A character vector.
}
\keyword{datasets}
