% Generated by roxygen2 (4.0.2): do not edit by hand
\name{splitdf}
\alias{splitdf}
\title{This will randomly split a data frame into training and test sets. I
  borrowed it from \href{http://gettinggeneticsdone.blogspot.com/2011/02/split-data-frame-into-testing-and.html}{here}.
  The caret packages has a similar facility using its createDataPartition()
  function. Perhaps even more appealing, there is sample.split() in caTools.}
\usage{
splitdf(dataframe, weight = 2/3, seed = NULL)
}
\arguments{
\item{dataframe}{This is the data frame you want to split.}

\item{weight}{The weights for the different classes.}

\item{seed}{An optional seed.}
}
\value{
A list with two members; the training set and the test set.
}
\description{
This will randomly split a data frame into training and test sets. I
  borrowed it from \href{http://gettinggeneticsdone.blogspot.com/2011/02/split-data-frame-into-testing-and.html}{here}.
  The caret packages has a similar facility using its createDataPartition()
  function. Perhaps even more appealing, there is sample.split() in caTools.
}

